# [Experiment] Use partition_by_hash in explicit_comms shuffle