Some experimental changes to the "shuffle group" stage of the `explicit_comms` shuffle code, motivated by a discovery by @wence- that the current `shuffle_group` code path may be 5x slower than `partition_by_hash`.

Note that I didn't see a clear performance difference on my local system with these changes, but I wasn't set up to run with UCX yet (so I'm sharing the code here anyway).