@wence- 2022-07-18
Thanks, I'll test on some scale-up runs, but this seems to make a (positive) difference for me:

## Trunk

```
Merge benchmark
--------------------------------------------------------------------------------
Backend                   | explicit-comms
Merge type                | gpu
Rows-per-chunk            | 100000000
Base-chunks               | 8
Other-chunks              | 8
Broadcast                 | default
Protocol                  | ucx
Device(s)                 | 0,1,2,3,4,5,6,7
RMM Pool                  | True
Frac-match                | 0.3
TCP                       | None
InfiniBand                | True
NVLink                    | True
Worker thread(s)          | 1
Data processed            | 23.84 GiB
Number of workers         | 8
================================================================================
Throughput                | 31.82 GiB/s +/- 153.91 MiB/s
Bandwidth                 | 587.51 MiB/s +/- 16.98 MiB/s
Wall clock                | 749.39 ms +/- 11.20 ms
================================================================================
```

## This branch

```
Merge benchmark
--------------------------------------------------------------------------------
Backend                   | explicit-comms
Merge type                | gpu
Rows-per-chunk            | 100000000
Base-chunks               | 8
Other-chunks              | 8
Broadcast                 | default
Protocol                  | ucx
Device(s)                 | 0,1,2,3,4,5,6,7
RMM Pool                  | True
Frac-match                | 0.3
TCP                       | None
InfiniBand                | True
NVLink                    | True
Worker thread(s)          | 1
Data processed            | 23.84 GiB
Number of workers         | 8
================================================================================
Throughput                | 36.75 GiB/s +/- 154.26 MiB/s
Bandwidth                 | 1.06 GiB/s +/- 63.02 MiB/s
Wall clock                | 648.80 ms +/- 8.41 ms
================================================================================```

The wallclock drops by 100ms which is twice the reduction in the cost of `shuffle_group` (which this change directly targets), but things are too dynamic for me to draw strong conclusions from that.