@madsbk 2022-08-01
Looks like a good approach to me